requirements.txt
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
opencv-python==4.10.0.84
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
scikit-image==0.24.0
pillow==11.0.0
einops==0.8.0